New lock with notification

+++ src/test/java/org/oba/jedis/extra/utils/cache/SimpleCacheLoaderTest.java
@@ -170,7 +170,7 @@ public class SimpleCacheLoaderTest {
 
         private synchronized void doWait() {
             try {
-                Thread.sleep(ThreadLocalRandom.current().nextInt(50));
+                Thread.sleep(ThreadLocalRandom.current().nextLong(50));
             } catch (InterruptedException e) {
                 // Eat this
             }
