+++ src/test/java/org/oba/jedis/extra/utils/cache/functional/FunctionalSimpleCacheWriterTest.java
@@ -428,7 +428,7 @@ public class FunctionalSimpleCacheWriterTest {
                 throw new IllegalStateException("Test error");
             }
             try {
-                Thread.sleep(ThreadLocalRandom.current().nextInt(10));
+                Thread.sleep(ThreadLocalRandom.current().nextLong(10L));
             } catch (InterruptedException e) {
                 // Eat this
             }
